test(ci): split the fork-read-failure cases into their own module
The previous commit's two new cases took
tests/ci/test_count_ratchet_fork_point.py to 496 lines against the 500-line
taste ceiling, so the next contributor to touch it would have raised the
count. Same move that module records having made out of
test_count_ratchet_against_real_git.py for the same reason.

The three fork-read-failure cases group cleanly: they were one case until
the previous commit, and what each now pins is that it blocks under its own
message and exit code. The direction cases stay behind.

`main_lowered_to_99`, `run_with_count`, `git_checked` and `git_stdout` move
into count_ratchet_git_harness.py rather than being copied, which its own
docstring already names as the thing that drifts.

267, 199 and 161 lines against the ceiling, and the same 50 tests pass.

Refs #5065
